Fix FIX2LONG addition
We cannot add to buffer_size because it is a Ruby VALUE object. By adding
one to buffer_size we turn it from a fixnum into something else because
fixnums use the bottom bit to determine that is a fixnum. Instead, we
should be incrementing the value after the FIX2LONG.

Co-Authored-By: Matt Valentine-House <matt@eightbitraptor.com>